requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fake-useragent==1.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.22.2
oauthlib==3.2.2
openpyxl==3.0.9
outcome==1.3.0.post0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
tzdata==2023.4
undetected-chromedriver==3.5.4
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
xlrd==2.0.1
zipp==3.17.0
